Demo of failing `nix flake check --no-build`
This is the smallest repro I can make of the issue I'm facing with
`nix flake check --no-build`.

Checking this flake with --no-build fails with:
error: path '/nix/store/rj6lbx32x9wdmzigvi55nbqmr7f7d8bx-known_hosts.drv' is not valid

Checking this flake without --no-build succeeds. Checking this flake
with --no-build also succeeds if the derivation was built and is
cached in the store.

Why I bumped into this: I'm running `nix flake check` in CI, and my CI is
on someone else's computer within a docker container that runs on
x86_64 hardware. As my flake contains nixosConfigurations for machines
of multiple architectures, this worker won't build them and this fails
`nix flake check` (even if I don't pass --all-systems!). And of course
CI workers start with an empty store.